refactor: extract date helpers to src/lib/date.ts and remove redundant props
- Move getTodayStr/getTomorrowStr/getWeekendStr/formatDateLabel/formatScheduleBadge
  into shared src/lib/date.ts to eliminate duplicate logic across TaskItem and SchedulingPicker
- Remove onScheduleTask/onUnscheduleTask from sub-task TaskItem renders
  (calendar button is hidden on isSubTask, so props were dead weight)
